+++ ft_printnbr.c
@@ -6,7 +6,7 @@
 /*   By: vabad-ro <vabad-ro@student.42madrid.com    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2026/01/28 13:51:22 by vabad-ro          #+#    #+#             */
-/*   Updated: 2026/01/29 17:04:52 by vabad-ro         ###   ########.fr       */
+/*   Updated: 2026/01/29 18:07:56 by vabad-ro         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -25,7 +25,7 @@ int	ft_printnbr(int n)
 		i++;
 	}
 	else
-		l = n;
+		l = (unsigned int) n;
 	if (l >= 10)
 		i += ft_printnbr(l / 10);
 	i += ft_printchr((l % 10) + '0');
